Move fetch GitHub logic in its own file
